reorganize command and ecs cli structure
