+++ docs/reference.md
@@ -1,0 +1,10 @@
+---
+title: CLI Reference
+---
+{% include reference.md %}
+
+* [sonic ecs exec]({% link _reference/sonic-ecs exec.md %})
+* [sonic ecs send]({% link _reference/sonic-ecs send.md %})
+* [sonic list]({% link _reference/sonic-list.md %})
+* [sonic run_command]({% link _reference/sonic-run_command.md %})
+* [sonic ssh]({% link _reference/sonic-ssh.md %})
