+++ lib/sonic/cli.rb
@@ -1,48 +1,24 @@
 require 'thor'
 
 module Sonic
-  class CLI < Command
+  class CLI < BaseCommand
+    desc "ecs SUBCOMMAND", "ecs subcommands"
+    long_desc Help.text(:ecs)
+    subcommand "ecs", Ecs
+
+    desc "command SUBCOMMAND", "command subcommands"
+    long_desc Help.text(:command)
+    subcommand "command", Command
+
     class_option :verbose, type: :boolean
     class_option :noop, type: :boolean
-    bastion_option = Proc.new {
-      option :bastion, desc: "Bastion jump host to use.  Defaults to no bastion server."
-    }
-    cluster_option = Proc.new {
-      option :cluster, desc: "ECS Cluster to use.  Default cluster is default"
-    }
-
     desc "ssh [IDENTIFER]", "ssh into a instance using identifier. identifer can be several things: instance id, ec2 tag, ECS service name, etc"
     long_desc Help.text(:ssh)
     method_option :keys, :aliases => '-i', :desc => "comma separated list of ssh private key paths"
     method_option :retry, :aliases => '-r', :type => :boolean, :desc => "keep retrying the server login until successful. Useful when on newly launched instances."
-    bastion_option.call
+    option :bastion, desc: "Bastion jump host to use.  Defaults to no bastion server."
     def ssh(identifier, *command)
       Ssh.new(identifier, options.merge(command: command)).run
-    end
-
-    desc "ecs_exec [ECS_SERVICE]", "docker exec into running docker container associated with the service on a container instance"
-    long_desc Help.text(:ecs_exec)
-    bastion_option.call
-    cluster_option.call
-    def ecs_exec(service, *command)
-      Docker.new(service, options.merge(command: command)).exec
-    end
-
-    # Cannot name the command run because that is a reserved Thor keyword :(
-    desc "ecs_run [ECS_SERVICE]", "docker run with the service on a container instance"
-    long_desc Help.text(:ecs_run)
-    bastion_option.call
-    cluster_option.call
-    def ecs_run(service, *command)
-      Docker.new(service, options.merge(command: command)).run
-    end
-
-    desc "run_command [FILTER] [COMMAND]", "runs command across fleet of servers via AWS Run Command"
-    long_desc Help.text(:run_command)
-    option :zero_warn, type: :boolean, default: true, desc: "Warns user when no instances found"
-    # filter - Filter ec2 instances by tag name or instance_ids separated by commas
-    def run_command(filter, *command)
-      RunCommand.new(command, options.merge(filter: filter)).execute
     end
 
     desc "list [FILTER]", "lists ec2 instances"
@@ -51,6 +27,23 @@ module Sonic
     # filter - Filter ec2 instances by tag name or instance_ids separated by commas
     def list(filter)
       List.new(options.merge(filter: filter)).run
+    end
+
+    desc "completion *PARAMS", "Prints words for auto-completion."
+    long_desc Help.text("completion")
+    def completion(*params)
+      Completer.new(CLI, *params).run
+    end
+
+    desc "completion_script", "Generates a script that can be eval to setup auto-completion."
+    long_desc Help.text("completion_script")
+    def completion_script
+      Completer::Script.generate
+    end
+
+    desc "version", "prints version"
+    def version
+      puts VERSION
     end
   end
 end
