+++ lib/sonic/command.rb
@@ -1,99 +1,11 @@
-require "thor"
-
-# Override thor's long_desc identation behavior
-# https://github.com/erikhuda/thor/issues/398
-class Thor
-  module Shell
-    class Basic
-      def print_wrapped(message, options = {})
-        message = "\n#{message}" unless message[0] == "\n"
-        stdout.puts message
-      end
-    end
-  end
-end
-
 module Sonic
-  class Command < Thor
-    desc "completion *PARAMS", "Prints words for auto-completion."
-    long_desc Help.text("completion")
-    def completion(*params)
-      Completer.new(CLI, *params).run
-    end
-
-    desc "completion_script", "Generates a script that can be eval to setup auto-completion."
-    long_desc Help.text("completion_script")
-    def completion_script
-      Completer::Script.generate
-    end
-
-    desc "version", "prints version"
-    def version
-      puts VERSION
-    end
-
-    class << self
-      def dispatch(m, args, options, config)
-        # Allow calling for help via:
-        #   sonic command help
-        #   sonic command -h
-        #   sonic command --help
-        #   sonic command -D
-        #
-        # as well thor's normal way:
-        #
-        #   sonic help command
-        help_flags = Thor::HELP_MAPPINGS + ["help"]
-        if args.length > 1 && !(args & help_flags).empty?
-          args -= help_flags
-          args.insert(-2, "help")
-        end
-
-        #   sonic version
-        #   sonic --version
-        #   sonic -v
-        version_flags = ["--version", "-v"]
-        if args.length == 1 && !(args & version_flags).empty?
-          args = ["version"]
-        end
-
-        super
-      end
-
-      # Override command_help to include the description at the top of the
-      # long_description.
-      def command_help(shell, command_name)
-        meth = normalize_command_name(command_name)
-        command = all_commands[meth]
-        alter_command_description(command)
-        super
-      end
-
-      def alter_command_description(command)
-        return unless command
-
-        # Add description to beginning of long_description
-        long_desc = if command.long_description
-            "#{command.description}\n\n#{command.long_description}"
-          else
-            command.description
-          end
-
-        # add reference url to end of the long_description
-        unless website.empty?
-          full_command = [command.ancestor_name, command.name].compact.join('-')
-          url = "#{website}/reference/sonic-#{full_command}"
-          long_desc += "\n\nAlso available at: #{url}"
-        end
-
-        command.long_description = long_desc
-      end
-      private :alter_command_description
-
-      # meant to be overriden
-      def website
-        ""
-      end
+  class Command < BaseCommand
+    desc "send [FILTER] [COMMAND]", "runs command across fleet of servers via AWS Run Command"
+    long_desc Help.text("command/send")
+    option :zero_warn, type: :boolean, default: true, desc: "Warns user when no instances found"
+    # filter - Filter ec2 instances by tag name or instance_ids separated by commas
+    def send(filter, *command)
+      RunCommand.new(command, options.merge(filter: filter)).execute
     end
   end
 end
