+++ lib/sonic/help/command/send.md
@@ -2,9 +2,9 @@ Run as a command across a list of servers. A filter must be provided.  The filte
 
 Examples:
 
-    $ sonic run_command hi-web-prod uptime
-    $ sonic run_command hi-web-prod,hi-worker-prod,hi-clock-prod uptime
-    $ sonic run_command i-030033c20c54bf149,i-030033c20c54bf150 uname -a
-    $ sonic run_command i-030033c20c54bf149 file://hello.sh
+    $ sonic command send hi-web-prod uptime
+    $ sonic command send hi-web-prod,hi-worker-prod,hi-clock-prod uptime
+    $ sonic command send i-030033c20c54bf149,i-030033c20c54bf150 uname -a
+    $ sonic command send i-030033c20c54bf149 file://hello.sh
 
 You cannot mix instance ids and tag names in the filter.
